game-flow: merge reader via libtrx
