+++ src/libtrx/include/libtrx/game/game_flow.h
@@ -2,5 +2,6 @@
 
 #include "./game_flow/common.h"
 #include "./game_flow/enum.h"
+#include "./game_flow/reader.h"
 #include "./game_flow/sequencer.h"
 #include "./game_flow/types.h"
