+++ src/tr1/game/game_flow.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "./game_flow/common.h"
-#include "./game_flow/reader.h"
 #include "./game_flow/sequencer.h"
 #include "./game_flow/vars.h"
+
+#include <libtrx/game/game_flow.h>
