+++ src/tr2/game/game_flow.h
@@ -4,3 +4,5 @@
 #include "game/game_flow/inventory.h"
 #include "game/game_flow/sequencer.h"
 #include "game/game_flow/vars.h"
+
+#include <libtrx/game/game_flow.h>
