--- src/tr2/game/game_flow/reader.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void GF_Load(const char *path);
