+++ src/tr2/meson.build
@@ -104,7 +104,6 @@ sources = [
   'game/game.c',
   'game/game_string.c',
   'game/game_flow/common.c',
-  'game/game_flow/reader.c',
   'game/game_flow/inventory.c',
   'game/game_flow/sequencer.c',
   'game/game_flow/sequencer_misc.c',
